requirements.txt
streamlit
pillow
google-generativeai
python-dotenv
opencv-python-headless
tqdm
